Introduce intermediary stage to allow proper resource recreation
Unfortunately the current setup still has the issue that prevents us from replacing the API-GW resources when the Swagger file changes.

Fortunately someone else has chimed in with their solution, and this seemed to work during my tests:
https://github.com/hashicorp/terraform/issues/6613#issuecomment-342237535

Also changed the output of the module to reflect custom domains.

(My test case was only about changing the description variable of the API Gateway module, which affects the Swagger file.)