Adds lambda module that uses code in S3
For months now many projects have been using Lambdas where the code is
located in s3 versus AWS's built in Lambda storage. There's two HUGE
benefits to this:

* we don't hit limits on the amount of code we're storing
* Terraform runs *far* faster as it isn't necessary to upload the same
  lambda zip file multiple times when all that changes is the handler
  module.

I wanted to built the S3 functionality into the existing lambda modules,
but that is sadly not possible due to the fact Terraform lacks a true
way to define "null" even though it must use such a mechanism
internally. When the "aws_lambda_function" resource is defined, it can
either set the `s3_bucket` and `s3_key` fields or the `filename` field,
but not both. To make a module that would support either approach, we'd
need to be able to pass `null` or something to the fields we aren't
using but `null` isn't going to be a thing until Terraform 0.12 is out.
Even then I'm not positive it will fix this problem.

I've been waiting for months for Terraform 0.12 to come out and during
that time I've copied the "s3" versions of the Lambda module to several
repos. Since it's such a drastic improvement I think we should just
give in and make a new module here instead.